requirements.txt
PyQt5==5.15.11
PyQtWebEngine==5.15.7
pyserial==3.5
pyinstaller==6.19.0
pyqtgraph==0.14.0
numpy==2.4.2
pytest==8.1.1
pytest-qt==4.4.0